mask-rcnn/requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.9.0
tensorflow-gpu==1.9.0
keras==2.1.3
opencv-python
h5py
imgaug
IPython[all]
ipykernel
pycocotools
lxml